+++ Src/FluentAssertions/Equivalency/IEquivalencyOptions.cs
@@ -6,9 +6,9 @@ using FluentAssertions.Equivalency.Tracing;
 namespace FluentAssertions.Equivalency;
 
 /// <summary>
-/// Provides the run-time details of the <see cref="EquivalencyAssertionOptions{TSubject}" /> class.
+/// Provides the run-time details of the <see cref="EquivalencyOptions{TExpectation}" /> class.
 /// </summary>
-public interface IEquivalencyAssertionOptions
+public interface IEquivalencyOptions
 {
     /// <summary>
     /// Gets an ordered collection of selection rules that define what members (e.g. properties or fields) are included.