+++ Src/FluentAssertions/Equivalency/IMemberMatchingRule.cs
@@ -27,5 +27,5 @@ public interface IMemberMatchingRule
     /// Returns the <see cref="IMember"/> of the property with which to compare the subject with, or <see langword="null"/>
     /// if no match was found.
     /// </returns>
-    IMember Match(IMember expectedMember, object subject, INode parent, IEquivalencyAssertionOptions options);
+    IMember Match(IMember expectedMember, object subject, INode parent, IEquivalencyOptions options);
 }